Improve media enocder test coverage

+++ src/av/src/multiplexencoder.cpp
@@ -230,21 +230,18 @@ void MultiplexEncoder::flush()
 
 EncoderOptions& MultiplexEncoder::options()
 {
-   
     return _options;
 }
 
 
 VideoEncoder* MultiplexEncoder::video()
 {
-   
     return _video;
 }
 
 
 AudioEncoder* MultiplexEncoder::audio()
 {
-   
     return _audio;
 }
 
@@ -289,11 +286,11 @@ bool MultiplexEncoder::updateStreamPts(AVStream* stream, int64_t* pts)
     if (*pts == AV_NOPTS_VALUE) {
         // Set a realtime pts value if not specified
         int64_t delta(av_gettime() - _formatCtx->start_time_realtime);
-        next = delta * (float)stream->time_base.den / (float)stream->time_base.num / AV_TIME_BASE;
+        next = delta * (double)stream->time_base.den / (double)stream->time_base.num / AV_TIME_BASE;
     } 
     else {
         // Convert from input microseconds to encoder stream time base
-        next = *pts * (float)stream->time_base.den / (float)stream->time_base.num / AV_TIME_BASE;
+        next = *pts * (double)stream->time_base.den / (double)stream->time_base.num / AV_TIME_BASE;
     }
 
     if (next == _pts) {
